Снова создаю обновленныый PointController
